+++ src/com/github/igotyou/FactoryMod/Factorys/ProductionFactory.java
@@ -19,6 +19,7 @@ import com.github.igotyou.FactoryMod.interfaces.Factory;
 import com.github.igotyou.FactoryMod.interfaces.Recipe;
 import com.github.igotyou.FactoryMod.listeners.RedstoneListener;
 import com.github.igotyou.FactoryMod.properties.ProductionProperties;
+import com.github.igotyou.FactoryMod.recipes.ProbabilisticEnchantment;
 import com.github.igotyou.FactoryMod.recipes.ProductionRecipe;
 import com.github.igotyou.FactoryMod.utility.InteractionResponse;
 import com.github.igotyou.FactoryMod.utility.InteractionResponse.InteractionResult;
@@ -29,19 +30,15 @@ import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 
-public class ProductionFactory extends FactoryObject implements Factory
+public class ProductionFactory extends BaseFactory
 {
-	public static final BlockFace[] REDSTONE_FACES = {BlockFace.NORTH, BlockFace.EAST, BlockFace.SOUTH, BlockFace.WEST, BlockFace.UP, BlockFace.DOWN};
 
 	private ProductionRecipe currentRecipe = null;//the recipe that is currently selected
 	private ProductionProperties productionFactoryProperties;//the properties of the production factory
-	private int currentProductionTimer = 0;//The "production timer", which trachs for how long the factory has been producing the selected recipe
-	private int currentEnergyTimer = 0;//Time since last energy consumption(if there's no lag, it's in seconds)
-	private int currentRecipeNumber = 0;//the array index of the current recipe
 	public static final FactoryType FACTORY_TYPE = FactoryType.PRODUCTION;//the factory's type
 	private List<ProductionRecipe> recipes;
-	private double currentRepair;
-	private long timeDisrepair;//The time at which the factory went into disrepair
+	private int currentRecipeNumber = 0;//the array index of the current recipe
+	
 	/**
 	 * Constructor
 	 */
@@ -52,8 +49,6 @@ public class ProductionFactory extends FactoryObject implements Factory
 		this.productionFactoryProperties = (ProductionProperties) factoryProperties;
 		this.recipes=new ArrayList<ProductionRecipe> (productionFactoryProperties.getRecipes());
 		this.setRecipeToNumber(0);
-		this.currentRepair=0.0;
-		this.timeDisrepair=3155692597470L;//Year 2070, default starting value
 	}
 
 	/**
@@ -63,228 +58,22 @@ public class ProductionFactory extends FactoryObject implements Factory
 			String subFactoryType, boolean active, int currentProductionTimer, int currentEnergyTimer,  List<ProductionRecipe> recipes,
 			int currentRecipeNumber,double currentMaintenance,long timeDisrepair)
 	{
-		super(factoryLocation, factoryInventoryLocation, factoryPowerSource, ProductionFactory.FACTORY_TYPE, subFactoryType);
+		super(factoryLocation, factoryInventoryLocation, factoryPowerSource, ProductionFactory.FACTORY_TYPE, active, subFactoryType, currentProductionTimer, currentEnergyTimer, currentMaintenance, timeDisrepair);
 		this.productionFactoryProperties = (ProductionProperties) factoryProperties;
-		this.active = active;
-		this.currentEnergyTimer = currentEnergyTimer;
-		this.currentProductionTimer = currentProductionTimer;
 		this.recipes=recipes;
 		this.setRecipeToNumber(currentRecipeNumber);
-		this.currentRepair=currentMaintenance;
-		this.timeDisrepair=timeDisrepair;
 	}
 	
-	/**
-	 * called by the manager each update cycle
-	 */
-	public void update() 
-	{
-		//if factory is turned on
-		if (active)
-		{
-			//if the materials required to produce the current recipe are in the factory inventory
-			if (currentRecipe.hasMaterials(getInventory()))
-			{
-				//if the factory has been working for less than the required time for the recipe
-				if (currentProductionTimer < currentRecipe.getProductionTime())
-				{
-					//if the factory power source inventory has enough fuel for at least 1 energyCycle
-					if (isFuelAvailable())
-					{
-						//if the time since fuel was last consumed is equal to how often fuel needs to be consumed
-						if (currentEnergyTimer == productionFactoryProperties.getEnergyTime()-1)
-						{
-							//remove one fuel.
-							productionFactoryProperties.getFuel().removeFrom(getPowerSourceInventory());
-							//0 seconds since last fuel consumption
-							currentEnergyTimer = 0;
-						}
-						//if we don't need to consume fuel, just increment the energy timer
-						else
-						{
-							currentEnergyTimer++;
-						}
-						//increment the production timer
-						currentProductionTimer ++;
-					}
-					//if there is no fuel Available turn off the factory
-					else
-					{
-						powerOff();
-					}
-				}
-				
-				//if the production timer has reached the recipes production time remove input from chest, and add output material
-				else if (currentProductionTimer >= currentRecipe.getProductionTime())
-				{
-					
-					//Remove inputs from chest
-					currentRecipe.getInputs().removeFrom(getInventory());
-					//Remove upgrade and replace it with its upgraded form
-					currentRecipe.getUpgrades().removeOneFrom(getInventory()).putIn(getInventory(),currentRecipe.getEnchantments());
-					//Adds outputs to chest with appropriate enchantments
-					currentRecipe.getOutputs().putIn(getInventory(),currentRecipe.getEnchantments());
-					//Adds new recipes to the factory
-					for (int i = 0; i < currentRecipe.getOutputRecipes().size();i++)
-					{
-						if(!recipes.contains(currentRecipe.getOutputRecipes().get(i)))
-						{
-							recipes.add(currentRecipe.getOutputRecipes().get(i));
-						}
-					}
-					//Repairs the factory
-					repair(currentRecipe.getRepairs().removeMaxFrom(getInventory(),(int)currentRepair));
-					//Remove currentRecipe if it only is meant to be used once
-					if(currentRecipe.getUseOnce())
-					{
-						recipes.remove(currentRecipe);
-						setRecipeToNumber(0);
-					}
-					currentProductionTimer = 0;
-					powerOff();
-				}
-			}	
-			else
-			{
-				powerOff();
-			}
-		}	
+	@Override
+	public boolean checkHasMaterials() {
+		return currentRecipe.hasMaterials(getInventory());
 	}
-	/**
-	 * Turns the factory on
-	 */
-	public void powerOn() 
-	{
-		//put active to true
-		active = true;
-		
-		// Set attached lever
-		if (FactoryModPlugin.LEVER_OUTPUT_ENABLED) {
-			setActivationLever(true);
-		}
-		
-		//lots of code to make the furnace light up, without loosing contents.
-		Furnace furnace = (Furnace) factoryPowerSourceLocation.getBlock().getState();
-		byte data = furnace.getData().getData();
-		ItemStack[] oldContents = furnace.getInventory().getContents();
-		furnace.getInventory().clear();
-		factoryPowerSourceLocation.getBlock().setType(Material.BURNING_FURNACE);
-		furnace = (Furnace) factoryPowerSourceLocation.getBlock().getState();
-		furnace.setRawData(data);
-		furnace.update();
-		furnace.getInventory().setContents(oldContents);
-		//reset the production timer
-		currentProductionTimer = 0;
+	
+	@Override
+	public boolean isRepairing() {
+		return currentRecipe.getRepairs().size()!=0;
 	}
-
-	private void setActivationLever(boolean state) {
-		Block lever = findActivationLever();
-		if (lever != null) {
-			setLever(lever, state);
-			shotGunUpdate(factoryPowerSourceLocation.getBlock());
-		}
-	}
-
-	/**
-	 * Turns the factory off.
-	 */
-	public void powerOff() 
-	{
-		if(active)
-		{
-			if (FactoryModPlugin.LEVER_OUTPUT_ENABLED) {
-				setActivationLever(false);
-			}
-			
-			//lots of code to make the furnace turn off, without loosing contents.
-			Furnace furnace = (Furnace) factoryPowerSourceLocation.getBlock().getState();
-			byte data = furnace.getData().getData();
-			ItemStack[] oldContents = furnace.getInventory().getContents();
-			furnace.getInventory().clear();
-			factoryPowerSourceLocation.getBlock().setType(Material.FURNACE);
-			furnace = (Furnace) factoryPowerSourceLocation.getBlock().getState();
-			furnace.setRawData(data);
-			furnace.update();
-			furnace.getInventory().setContents(oldContents);
-			
-			//put active to false
-			active = false;
-			//reset the production timer
-			currentProductionTimer = 0;
-		}
-	}
-
-	/**
-	 * Returns either a success or error message.
-	 * Called by the blockListener when a player left clicks the powerSourceLocation with the InteractionMaterial
-	 */
-	public List<InteractionResponse> togglePower() 
-	{
-		List<InteractionResponse> response=new ArrayList<InteractionResponse>();
-		//if the factory is turned off
-		if (!active)
-		{
-			//if the factory isn't broken or the current recipe can repair it
-			if(!isBroken()||currentRecipe.getRepairs().size()!=0)
-			{
-				//is there fuel enough for at least once energy cycle?
-				if (isFuelAvailable())
-				{
-					//are there enough materials for the current recipe in the chest?
-					if (currentRecipe.hasMaterials(getInventory()))
-					{
-						//turn the factory on
-						powerOn();
-						//return a success message
-						response.add(new InteractionResponse(InteractionResult.SUCCESS, "Factory activated!"));
-						return response;
-					}
-					//there are not enough materials for the recipe!
-					else
-					{
-						//return a failure message, containing which materials are needed for the recipe
-						//[Requires the following: Amount Name, Amount Name.]
-						//[Requires one of the following: Amount Name, Amount Name.]
-						ItemList<NamedItemStack> needAll=new ItemList<NamedItemStack>();
-						needAll.addAll(currentRecipe.getInputs().getDifference(getInventory()));
-						needAll.addAll(currentRecipe.getRepairs().getDifference(getInventory()));
-						if(!needAll.isEmpty())
-						{
-							response.add(new InteractionResponse(InteractionResult.FAILURE,"You need all of the following: "+needAll.toString()+"."));
-						}
-						if(!currentRecipe.getUpgrades().oneIn(getInventory()))
-						{
-							response.add(new InteractionResponse(InteractionResult.FAILURE,"You need one of the following: "+currentRecipe.getUpgrades().toString()+"."));
-
-						}
-						return response;
-					}
-				}
-				//if there isn't enough fuel for at least one energy cycle
-				else
-				{
-					//return a error message
-					int multiplesRequired=(int)Math.ceil(currentRecipe.getProductionTime()/(double)getProductionFactoryProperties().getEnergyTime());
-					response.add(new InteractionResponse(InteractionResult.FAILURE, "Factory is missing fuel! ("+getProductionFactoryProperties().getFuel().getMultiple(multiplesRequired).toString()+")"));
-					return response;
-				}
-			}
-			else
-			{
-				response.add(new InteractionResponse(InteractionResult.FAILURE, "Factory is in disrepair!"));
-				return response;
-			}			
-		}
-		//if the factory is on already
-		else
-		{
-			//turn the factory off
-			powerOff();
-			//return success message
-			response.add(new InteractionResponse(InteractionResult.FAILURE, "Factory has been deactivated!"));
-			return response;
-		}
-	}
+	
 	
 	/**
 	 * Returns either a success or error message.
@@ -337,6 +126,11 @@ public class ProductionFactory extends FactoryObject implements Factory
 		return responses;
 	}
 	
+	@Override
+	public ItemList<NamedItemStack> getFuel() {
+		return productionFactoryProperties.getFuel();
+	}
+
 	/**
 	 * Sets the factories current recipe.
 	 * @param newRecipe the desired recipe
@@ -366,110 +160,6 @@ public class ProductionFactory extends FactoryObject implements Factory
 			currentRecipeNumber=0;
 		}
 	}
-
-	/**
-	 * returns the Location of the central block of the factory
-	 */
-	public Location getCenterLocation() 
-	{
-		return factoryLocation;
-	}
-
-	/**
-	 * returns the Location of the factory Inventory
-	 */
-	public Location getInventoryLocation() 
-	{
-		return factoryInventoryLocation;
-	}
-
-	/**
-	 * returns the Location of the factory power source
-	 */
-	public Location getPowerSourceLocation() 
-	{
-		return factoryPowerSourceLocation;
-	}
-	
-	/**
-	 * Checks if there is enough fuel Available for atleast once energy cycle
-	 * @return true if there is enough fuel, false otherwise
-	 */
-	public boolean isFuelAvailable()
-	{
-		return productionFactoryProperties.getFuel().allIn(getPowerSourceInventory());
-	}
-
-	/**
-	 * Called by the block listener when the player(or a entity) destroys the fatory
-	 * Drops the build materials if the config says it shouls
-	 */
-	public void destroy(Location destroyLocation)
-	{
-		powerOff();
-		currentRepair=getProductionFactoryProperties().getRepair();
-		timeDisrepair=System.currentTimeMillis();
-	}
-	/*
-	 * Repairs the factory 
-	 */
-	private void repair(int amountRepaired)
-	{
-		currentRepair-=amountRepaired;
-		if(currentRepair<0)
-		{
-			currentRepair=0;
-		}
-		if(currentRepair<getProductionFactoryProperties().getRepair())
-		{
-			timeDisrepair=3155692597470L;//Year 2070, default starting value
-		}
-	}
-
-	/**
-	 * Degrades the factory
-	 */
-	public void updateRepair(double percent)
-	{
-		int totalRepair=getProductionFactoryProperties().getRepair();
-		currentRepair+=totalRepair*percent;
-		if(currentRepair>=totalRepair)
-		{
-			currentRepair=totalRepair;
-			long currentTime=System.currentTimeMillis();
-			if (currentTime<timeDisrepair&&getProductionFactoryProperties().getRepair()!=0)
-			{
-				timeDisrepair=currentTime;
-			}	
-		}
-	}
-	public double getCurrentRepair()
-	{
-		return currentRepair;
-	}
-	/**
-	 * Checks that a factory hasn't degraded too much
-	 */
-	public boolean isBroken()
-	{
-		return currentRepair>=getProductionFactoryProperties().getRepair()&&getProductionFactoryProperties().getRepair()!=0;
-	}
-	/**
-	 * Returns the production timer
-	 */
-	public int getProductionTimer()
-	{
-		return currentProductionTimer;
-	}
-	
-	/**
-	 * Returns the energy timer
-	 */
-	public int getEnergyTimer()
-	{
-		return currentEnergyTimer;
-	}
-
 	/**
 	 * Returns the currentRecipe
 	 */
@@ -520,9 +210,9 @@ public class ProductionFactory extends FactoryObject implements Factory
 			responses.add(new InteractionResponse(InteractionResult.SUCCESS,"Upgrades: "+currentRecipe.getUpgrades().toString()+"."));
 		}
 		//[Outputs: amount Name, amount Name.]
-		if(!currentRecipe.getOutputs().isEmpty())
+		if(!getOutputs().isEmpty())
 		{
-			responses.add(new InteractionResponse(InteractionResult.SUCCESS,"Output: "+currentRecipe.getOutputs().toString()+"."));
+			responses.add(new InteractionResponse(InteractionResult.SUCCESS,"Output: "+getOutputs().toString()+"."));
 		}
 		//[Will repair XX% of the factory]
 		if(!currentRecipe.getRepairs().isEmpty()&&getProductionFactoryProperties().getRepair()!=0)
@@ -551,86 +241,59 @@ public class ProductionFactory extends FactoryObject implements Factory
 		return responses;
 	}
 	
-	/**
-	 * returns the date at which the factory went into disrepair
-	 */
-	public long getTimeDisrepair()
-	{
-		return timeDisrepair;
-	}
-	
-	public Block findActivationLever() {
-    	Block block = getPowerSourceLocation().getBlock();
-    	return findAttachedLever(block);
-	}
-    
-    public Block findAttachedLever(Block block) {
-		// Check sides for attached lever - required for automation
-		Block lever = null;
-		for (BlockFace face : REDSTONE_FACES) {
-			lever = block.getRelative(face);
-			if (lever.getType() == Material.LEVER) {
-			    BlockFace attached = getAttachedFace(lever);
-			    if (attached != null && attached == face) {
-					return lever;
-			    }
+	protected void recipeFinished() {
+		//Remove upgrade and replace it with its upgraded form
+		currentRecipe.getUpgrades().removeOneFrom(getInventory()).putIn(getInventory(),currentRecipe.getEnchantments());
+		//Adds new recipes to the factory
+
+		for (int i = 0; i < currentRecipe.getOutputRecipes().size();i++)
+		{
+			if(!recipes.contains(currentRecipe.getOutputRecipes().get(i)))
+			{
+				recipes.add(currentRecipe.getOutputRecipes().get(i));
 			}
 		}
 		
-		return null;
-    }
-    
-    private static BlockFace getAttachedFace(Block lever) {
-    	BlockState state = lever.getState();
-    	MaterialData md = state.getData();
-    	if (md instanceof Attachable) {
-    		BlockFace face = ((Attachable) md).getAttachedFace();
-    		return face.getOppositeFace();
-    	} else {
-    		return null;
-    	}
-    }
-    
-    private void shotGunUpdate(Block block) {
-    	for (BlockFace direction : REDSTONE_FACES) {
-    		Block neighbour = block.getRelative(direction);
-    		
-    	}
-    }
-    
-    
-	/**
-	* Sets the toggled state of a single lever<br>
-	* <b>No Lever type check is performed</b>
-	*
-	* @param lever block
-	* @param down state to set to
-	*/
-	private static void setLever(org.bukkit.block.Block lever, boolean down) {
-		if (lever.getType() != Material.LEVER) {
-			return;
+		//Remove currentRecipe if it only is meant to be used once
+		if(currentRecipe.getUseOnce())
+		{
+			recipes.remove(currentRecipe);
+			setRecipeToNumber(0);
 		}
-		
-		byte data = lever.getData();
-		int newData;
-		if (down) {
-			newData = data | 0x8;
-		} else {
-			newData = data & 0x7;
-		}
-		if (newData != data) {
-			// CraftBukkit start - Redstone event for lever
-			int old = !down ? 1 : 0;
-			int current = down ? 1 : 0;
-			BlockRedstoneEvent eventRedstone = new BlockRedstoneEvent(lever, old, current);
-			Bukkit.getServer().getPluginManager().callEvent(eventRedstone);
-			if ((eventRedstone.getNewCurrent() > 0) != down) {
-				return;
-			}
-			// CraftBukkit end
-			lever.setData((byte) newData, true);
-			lever.getState().update();
-			Block attached = lever.getRelative(((Attachable) lever.getState().getData()).getAttachedFace());
-		}
+	}
+
+	@Override
+	public ItemList<NamedItemStack> getInputs() {
+		return currentRecipe.getInputs();
+	}
+
+	@Override
+	public ItemList<NamedItemStack> getOutputs() {
+		return currentRecipe.getOutputs();
+	}
+
+	@Override
+	public ItemList<NamedItemStack> getRepairs() {
+		return currentRecipe.getRepairs();
+	}
+
+	@Override
+	public List<ProbabilisticEnchantment> getEnchantments() {
+		return currentRecipe.getEnchantments();
+	}
+
+	@Override
+	public double getEnergyTime() {
+		return productionFactoryProperties.getEnergyTime();
+	}
+
+	@Override
+	public double getProductionTime() {
+		return currentRecipe.getProductionTime();
+	}
+
+	@Override
+	public int getMaxRepair() {
+		return productionFactoryProperties.getRepair();
 	}
 }
